Combined files for simplicity and renamed addAndDelete to courseManager

+++ src/courseManager.py
@@ -1,6 +1,8 @@
 import locale
+import os
+import shuttil
 
-class addAndDelete:
+class courseManager:
 	parent = None
 	
 	##to get a section for a course
@@ -16,6 +18,8 @@ class addAndDelete:
 		self.parent = parent
 	#}
 	
+''' ----------------------------- Section of code that is used for creating and deleting courses -------------------------------------------'''	
+	
 	##creates a new course directory in the instructors folder
 	##path is the path to where the course directory is to be created
 	##courseName is the name of the new course
@@ -25,8 +29,13 @@ class addAndDelete:
 		addFolder(newCoursePath)
 		
 		courseConfigFile = newCoursePath + "course.config" ##creates the course config file
-		if courseConfigFile == False: #{
-			
+		
+		##create the course config file
+		
+		
+		##if the course config file is not created False is returned to indicate an error
+		if courseConfigFile == False: #{ 
+			return False
 		#}
 			
 		##creates the course config file and updates the global config
@@ -38,20 +47,31 @@ class addAndDelete:
 	##deletes a course directory in the instructor's directory
 	##path is the path to the course directory
 	##courseName is the name of the course to be removed
-	def deleteCourse(path, courseName): #{
-		parent.ConfigParser.removeCourse(GLOBAL_PATH, courseName) ##removes the course from the global config file
+	def deleteCourse(courseName): #{
+		path = parent.ConfigParser.get_setting(GLOBAL_PATH, courseName, "course_path")
+	
+		check = parent.ConfigParser.removeCourse(GLOBAL_PATH, courseName) ##removes the course from the global config file
+		
+		if check == False: #{
+			return False
+		#}
 		
 		deleteFolder(path) ##deletes the course and all assignments under it
 		
 		return True
 	#}
+	
+	
+''' ----------------------------- Section of code that is used for creating and deleting assignments --------------------------------------'''	
 
 	##creates a new assignment directory inside the course directory
 	##assignmentName is the name of the new directory
 	##path is the path to the course directory, assignmentName is the name of the assignment, dueDate is the day the assignment is dueDate
 	##team identifies if the assignment is a team assignment, maxSubmissions are the total number of submissions allowed, lateDays are the 
 	##number of days allowed for late submission
-	def addAssignment(path, assignmentName, dueDate, team, maxSubmissions, lateDays): #{
+	def addAssignment(courseName, assignmentName, dueDate, team, maxSubmissions, lateDays): #{
+		path = parent.ConfigParser.get_setting(GLOBAL_PATH, courseName, "course_path")
+		
 		addFolder(path) ##creates a new folder for the assignment
 		
 		assignmentConfigFile = path + "assignment.config"
@@ -100,12 +120,59 @@ class addAndDelete:
 		return True
 	#}
 	
+''' ----------------------------------- Section of code that is used for grading -------------------------------------------------------'''	
+
+	##Gives a grade to the student
+	##path is the path to the student's directory in an assignment
+	##gradeRecieved is the grade recieved for the assignment
+	def enterGrade(path, gradeRecieved): #{
+		gradeFile = path + "grade.txt"
+		
+		file grade = open(gradeFile, "w")
+		
+		grade.write("Grade Recieved: " + gradeRecieved)
+		
+		grade.close()
+	#}
 	
+''' ----------------------------------- Section of code that is used for creating folders --------------------------------------------------'''	
+	#x is the path including the new directory name
+	def addFolder(x): #{
+		#NOTE!-------------------------------------------------------------------------------------
+		# mkdir has another parameter that sets permissions for the new directory
+		#!-----------------------------------------------------------------------------------------
+		os.mkdir(x); #a new directory is made
+	#}
+
+	def deleteFolder(x): #{
+		os.rmtree(x) ##removes the directory and all directories and files inside it
+	#}
 	
+'''---------------------------------- Code that is used for testing ------------------------------------------------------------------'''
+	##tests the addCourse function
+	def testAddCourse() #{
+		
+	#}
 	
+	##tests the deleteCourse function
+	def testDeleteCourse() #{
+		
+	#}
+
+	##tests the addAssignment function
+	def testAddAssignment() #{
+		
+	#}
+
+	##tests the deleteAssignment function
+	def testDeleteAssignment() #{
+		
+	#}
 	
-	
-	
+	##main method
+	def main(): #{
+		
+	#}
